@ikawrakow 2025-01-29
Here is how much time is being spent in the various matrix multiplications in the attention part when processing a prompt of 8192 tokens:

| result tensor |  time (s) |
| ------------: | ---------: |
| kq                  |  4.116 |
| kqv                 |  2.372 |
| kqv_out             |  0.458 |
| kv                  |  0.253 |
| kv_pe_compresseed   |  0.219 |
| q                   |  0.687 |
| total               |  8.107 |

And here is with this PR:

| result tensor |  time (s) |
| ------------: | ---------: |
| kq_nope             |  8.343 |
| kq_pe               |  2.495 |
| kqv                 |  0.401 |
| kqv_compressed      |  7.120 |
| kqv_out             |  0.473 |
| kv_pe_compresseed   |  0.224 |
| q                   |  0.693 |
| q_nope2             |  0.240 |
| total               |  19.989 |

I.e., attention is 2.5X slower with the PR. In addition, I'm finding that on the main branch `0.114` seconds are spent in `GGML_OP_ADD` operations, and `0.194` seconds for `GGML_OP_CONT`. In this PR  `3.320` seconds go into `GGML_OP_ADD`, and `2.701` seconds into `GGML_OP_CONT` (basically making copies). For reference, total processing time is `27.73` seconds on main and `45.47` seconds with the PR.

Maybe this can be useful when trying to optimize.