@saood06 2025-01-29
>This hurts prompt processing (a.k.a prefill) speed very significantly.
>[...]
>I think we need to either try to understand why the attention part is so much slower when processing batches of tokens and fix it, or simply wait for @fairydreaming to fix their PR.

Changed to draft. PP does seem to have regressions, I'll have direct comparisons against old version soon, generating an iq4_k_r4 quant now (PP in main for me was 11.5 t/s for iq4_k and  9.8 t/s for iq4_k_r4 at pp512,  9.22 t/s at PP1024 for IQ4_K).

>Maybe this can be useful when trying to optimize.

Thank you for the op time breakdown. 

I was drawn in to this PR for the TG benefits, it should have also been a draft for the reason that it would mean GGUF's wouldn't be cross compatible, as this is also a draft in llama.cpp. I just want to have it here because it does optimize for a workload where TG dominates, and R1 as a reasoning model it often does.