@ikawrakow 2025-01-29
@fairydreaming 

Great!

Btw, I observe that `attn_kv_b.weight` is still present in the model. Is it needed, given that we now have `attn_k_b.weight` and `attn_v_b.weight` ?