@saood06 2025-01-30
I ran batched-bench at batch size 1 with TG at 32 at various PP to show PP performance and TG performance at different context lengths. Batched-bench numbers are noisy because they do not use repetitions like llama-bench and this model on this machine seems to have some variance, but all data is shown after dropping the cache's and running the model until it is fully in the page cache.

IQ4_K_R4 with this PR:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   22.569 |     5.67 |   10.237 |     3.13 |   32.806 |     4.88 |
|   256 |     32 |    1 |    288 |   38.648 |     6.62 |   10.699 |     2.99 |   49.347 |     5.84 |
|   512 |     32 |    1 |    544 |   76.447 |     6.70 |   10.793 |     2.96 |   87.240 |     6.24 |
|  1024 |     32 |    1 |   1056 |  144.100 |     7.11 |   10.788 |     2.97 |  154.888 |     6.82 |
|  2048 |     32 |    1 |   2080 |  312.306 |     6.56 |   12.624 |     2.53 |  324.930 |     6.40 |
|  4096 |     32 |    1 |   4128 |  745.760 |     5.49 |   12.929 |     2.48 |  758.688 |     5.44 |
|  8192 |     32 |    1 |   8224 | 2023.859 |     4.05 |   16.017 |     2.00 | 2039.877 |     4.03 |

IQ4_K_R4 on main:
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   20.958 |     6.11 |   10.999 |     2.91 |   31.956 |     5.01 |
|   256 |     32 |    1 |    288 |   38.777 |     6.60 |   11.780 |     2.72 |   50.558 |     5.70 |
|   512 |     32 |    1 |    544 |   63.574 |     8.05 |   12.474 |     2.57 |   76.047 |     7.15 |
|  1024 |     32 |    1 |   1056 |  118.630 |     8.63 |   14.462 |     2.21 |  133.092 |     7.93 |
|  2048 |     32 |    1 |   2080 |  258.999 |     7.91 |   18.241 |     1.75 |  277.239 |     7.50 |
|  4096 |     32 |    1 |   4128 |  574.593 |     7.13 |   26.023 |     1.23 |  600.616 |     6.87 |
|  8192 |     32 |    1 |   8224 | 1391.722 |     5.89 |   43.056 |     0.74 | 1434.778 |     5.73 |


Looking at the 8K context results, PP does drop from 5.89 to 4.05, but TG jumps from 0.74 to 2.00. At q8_0 (results below) PP again drops 6.06 to 4.03, but TG benefits going from 0.99 to 1.94. I would test/run this model at even higher context, but I would either need a smaller quant or to use RPC (for reference the KV cache at n_ctx of 8224 is 40,233.55 MiB)

<details>
  <summary>Expand to see more runs with q8_0 and q6_0 K cache tested as well</summary>

  PR with q6_0 K cache:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   14.948 |     8.56 |   10.498 |     3.05 |   25.446 |     6.29 |
|   256 |     32 |    1 |    288 |   35.061 |     7.30 |   10.430 |     3.07 |   45.491 |     6.33 |
|   512 |     32 |    1 |    544 |   69.842 |     7.33 |   10.936 |     2.93 |   80.778 |     6.73 |
|  1024 |     32 |    1 |   1056 |  142.141 |     7.20 |   11.083 |     2.89 |  153.224 |     6.89 |
|  2048 |     32 |    1 |   2080 |  313.431 |     6.53 |   11.415 |     2.80 |  324.846 |     6.40 |
|  4096 |     32 |    1 |   4128 |  763.385 |     5.37 |   12.964 |     2.47 |  776.349 |     5.32 |
|  8192 |     32 |    1 |   8224 | 2076.578 |     3.94 |   16.371 |     1.95 | 2092.948 |     3.93 |


  PR with q8_0 K cache:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   15.804 |     8.10 |   10.288 |     3.11 |   26.092 |     6.13 |
|   256 |     32 |    1 |    288 |   34.806 |     7.35 |   10.436 |     3.07 |   45.242 |     6.37 |
|   512 |     32 |    1 |    544 |   69.839 |     7.33 |   10.597 |     3.02 |   80.437 |     6.76 |
|  1024 |     32 |    1 |   1056 |  141.519 |     7.24 |   10.909 |     2.93 |  152.428 |     6.93 |
|  2048 |     32 |    1 |   2080 |  310.669 |     6.59 |   11.430 |     2.80 |  322.099 |     6.46 |
|  4096 |     32 |    1 |   4128 |  751.935 |     5.45 |   12.970 |     2.47 |  764.905 |     5.40 |
|  8192 |     32 |    1 |   8224 | 2031.924 |     4.03 |   16.499 |     1.94 | 2048.424 |     4.01 |
  
  Second run of PR without K cache quantization:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   20.898 |     6.12 |   10.378 |     3.08 |   31.276 |     5.12 |
|   256 |     32 |    1 |    288 |   40.503 |     6.32 |   10.407 |     3.07 |   50.910 |     5.66 |
|   512 |     32 |    1 |    544 |   70.978 |     7.21 |   10.629 |     3.01 |   81.607 |     6.67 |
|  1024 |     32 |    1 |   1056 |  144.713 |     7.08 |   10.879 |     2.94 |  155.592 |     6.79 |
|  2048 |     32 |    1 |   2080 |  311.658 |     6.57 |   11.718 |     2.73 |  323.376 |     6.43 |
|  4096 |     32 |    1 |   4128 |  754.120 |     5.43 |   12.996 |     2.46 |  767.116 |     5.38 |
|  8192 |     32 |    1 |   8224 | 2037.022 |     4.02 |   16.437 |     1.95 | 2053.458 |     4.00 |

  main with q6_0 K cache:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   18.503 |     6.92 |   10.480 |     3.05 |   28.983 |     5.52 |
|   256 |     32 |    1 |    288 |   31.320 |     8.17 |   10.858 |     2.95 |   42.178 |     6.83 |
|   512 |     32 |    1 |    544 |   57.909 |     8.84 |   11.459 |     2.79 |   69.368 |     7.84 |
|  1024 |     32 |    1 |   1056 |  118.199 |     8.66 |   12.679 |     2.52 |  130.878 |     8.07 |
|  2048 |     32 |    1 |   2080 |  250.592 |     8.17 |   15.486 |     2.07 |  266.078 |     7.82 |
|  4096 |     32 |    1 |   4128 |  541.938 |     7.56 |   20.315 |     1.58 |  562.253 |     7.34 |
|  8192 |     32 |    1 |   8224 | 1353.169 |     6.05 |   30.144 |     1.06 | 1383.313 |     5.95 |




  main with q8_0 K cache:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |     32 |    1 |    160 |   16.825 |     7.61 |   10.586 |     3.02 |   27.411 |     5.84 |
|   256 |     32 |    1 |    288 |   33.362 |     7.67 |   10.894 |     2.94 |   44.255 |     6.51 |
|   512 |     32 |    1 |    544 |   54.048 |     9.47 |   11.869 |     2.70 |   65.917 |     8.25 |
|  1024 |     32 |    1 |   1056 |  109.381 |     9.36 |   13.128 |     2.44 |  122.509 |     8.62 |
|  2048 |     32 |    1 |   2080 |  238.006 |     8.60 |   15.567 |     2.06 |  253.574 |     8.20 |
|  4096 |     32 |    1 |   4128 |  553.239 |     7.40 |   21.099 |     1.52 |  574.339 |     7.19 |
|  8192 |     32 |    1 |   8224 | 1351.138 |     6.06 |   32.240 |     0.99 | 1383.377 |     5.94 |



</details>

>I think one should make Flash Attention work with different K and V head sizes. 

If that happened it would also have the benefit of allowing V cache quantization (not sure why FA is needed for that), which this model could really benefit from in it's current implementation which uses the space of MHA. A proper MLA implementation would take up far less space. 

>I'm playing with Deepseek-Lite and I'm finding that the CUDA performance is pretty bad

Other people have reported poor performance even for the larger Deepseek models with TG at 10-14 t/s (although with an IQ1 based quant) even fully offloaded with datacenter GPU's, and around the same performance for a 192GB Mac.

>So, I guess, your Epyc system wipes the floor with any GPU setup using partial GPU offload of Deepseek-R1.

Partial offload is reported benefited by this: https://github.com/ggerganov/llama.cpp/pull/11397 and it is something I plan to test/use.